refactor(openviking): split plugin into focused modules
